+++ src/components/icons/close-icon.tsx
