+++ src/components/icons/diamond-icon.tsx
