+++ src/components/icons/income-icon.tsx
