+++ src/components/icons/index.ts
@@ -1,18 +1,18 @@
-import HomeIcon from "./home_icon";
-import ShopIcon from "./shop_icon";
-import PromoteIcon from "./promote_icon";
-import DiamondIcon from "./diamond_icon";
-import ChevronDownIcon from "./chevron_down";
-import ChevronUpIcon from "./chevron_up";
-import IncomeIcon from "./income_icon";
-import UserIcon from "./user_icon";
-import SunIcon from "./sun_icon";
-import MoonIcon from "./moon_icon";
-import PlusIcon from "./plus_icon";
-import NotificationIcon from "./notification_icon";
-import MessageIcon from "./message_icon";
-import SearchIcon from "./search_icon";
-import QuestionIcon from "./question_icon";
-import MenuIcon from "./menu_icon";
-import CloseIcon from "./close_icon";
+import HomeIcon from "./home-icon";
+import ShopIcon from "./shop-icon";
+import PromoteIcon from "./promote-icon";
+import DiamondIcon from "./diamond-icon";
+import ChevronDownIcon from "./chevron-down";
+import ChevronUpIcon from "./chevron-up";
+import IncomeIcon from "./income-icon";
+import UserIcon from "./user-icon";
+import SunIcon from "./sun-icon";
+import MoonIcon from "./moon-icon";
+import PlusIcon from "./plus-icon";
+import NotificationIcon from "./notification-icon";
+import MessageIcon from "./message-icon";
+import SearchIcon from "./search-icon";
+import QuestionIcon from "./question-icon";
+import MenuIcon from "./menu-icon";
+import CloseIcon from "./close-icon";
 export {HomeIcon,PlusIcon,MenuIcon,CloseIcon, ShopIcon,UserIcon,MessageIcon,NotificationIcon,DiamondIcon,SearchIcon,ChevronDownIcon,IncomeIcon,ChevronUpIcon,PromoteIcon, QuestionIcon, SunIcon, MoonIcon };
