+++ src/components/icons/plus-icon.tsx
