+++ src/components/icons/promote-icon.tsx
