+++ src/components/icons/shop-icon.tsx
