+++ src/components/vendor-mobile-sidebar/index.tsx
@@ -1,39 +1,43 @@
-import React, { useState } from 'react'
-import { ChevronDownIcon, DiamondIcon, HomeIcon, IncomeIcon, MoonIcon, PromoteIcon, QuestionIcon, ShopIcon, SunIcon, UserIcon } from '../icons'
+import { useState } from 'react'
+import { ChevronDownIcon, CloseIcon, DiamondIcon, HomeIcon, IncomeIcon, PromoteIcon, QuestionIcon, ShopIcon, UserIcon } from '../icons'
 import { ListItem } from '../buttons'
-import ToggleLightDarkMode from '../buttons/toggle_light_dark_mode'
+import ToggleLightDarkMode from '../buttons/toggle-light-dark-mode'
 
-function DashboardSideBar() {
+function DashboardMobileSideBar({onClick}:any) {
     const [selectedItem, setSelected] = useState("home")
     return (
-        <div className='sm:flex h-full bg-white lg:min-w-[340px] hidden lg:max-w-[340px] p-6 gap-1 flex-col'>
-            <img src='logo.svg' className='w-14' />
-            <ListItem onClick={() => {
+        <div className='flex h-screen w-full  bg-white sm:min-w-[340px] absolute top-0 left-0 z-10 sm:max-w-[340px] p-6 gap-1 flex-col'>
+            <div className='flex w-full justify-between' onClick={()=>{
+                onClick()
+            }}>
+                <CloseIcon className='sm:hidden flex mt-1 w-14' />
+                <img src='logo.svg' className='w-9' /></div>
+            <ListItem isMobile={true} onClick={() => {
                 setSelected("home")
-            }} selectedItem={selectedItem} item='home' text='Home' className='font-[600] text-gray-600 mt-9' iconLeft={<HomeIcon />} />
-            <ListItem onClick={() => {
+            }} selectedItem={selectedItem} item='home' text='Home' className='font-[600] !flex text-gray-600 mt-6' iconLeft={<HomeIcon />} />
+            <ListItem isMobile={true} onClick={() => {
                 setSelected("products")
             }} selectedItem={selectedItem} item='products' text='Products' iconRight={<ChevronDownIcon className='' />} className='font-[600] text-gray-600' iconLeft={<DiamondIcon />} />
-            <ListItem onClick={() => {
+            <ListItem isMobile={true} onClick={() => {
                 setSelected("customers")
             }} selectedItem={selectedItem} item='customers' text='Customers' iconRight={<ChevronDownIcon className='' />} className='font-[600] text-gray-600' iconLeft={<UserIcon />} />
-            <ListItem onClick={() => {
+            <ListItem isMobile={true} onClick={() => {
                 setSelected("shops")
             }} selectedItem={selectedItem} item='shops' text='Shop' className='font-[600] text-gray-600' iconLeft={<ShopIcon />} />
-            <ListItem onClick={() => {
+            <ListItem isMobile={true} onClick={() => {
                 setSelected("income")
             }} selectedItem={selectedItem} item='income' text='Income' iconRight={<ChevronDownIcon className='' />} className='font-[600] text-gray-600' iconLeft={<IncomeIcon />} />
-            <ListItem onClick={() => {
+            <ListItem isMobile={true} onClick={() => {
                 setSelected("promote")
             }} selectedItem={selectedItem} item='promote' text='Promote' className='font-[600] text-gray-600' iconLeft={<PromoteIcon />} />
             <div className='mt-auto flex pt-2 flex-col gap-2 border-t'>
-                <ListItem onClick={() => {
+                <ListItem isMobile={true} onClick={() => {
                     setSelected("help")
                 }} selectedItem={selectedItem} isDropdown={false} item='help' text='Help & getting started' className='font-[600] text-gray-600' iconLeft={<QuestionIcon />} iconRight={<div className='w-6 h-6 font-bold flex items-center justify-center rounded-lg bg-purple-300'>8</div>} />
-                <ToggleLightDarkMode isSmallScreen={false} />
+                <ToggleLightDarkMode isSmallScreen={true} />
             </div>
         </div>
     )
 }
 
-export default DashboardSideBar
+export default DashboardMobileSideBar
